pyproject.toml
[project]
name = "cv-dip-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "python-pptx>=1.0.2",
]
